update + move get_solver_ids
