+++ tuna/helper.py
@@ -34,6 +34,7 @@ from sqlalchemy.orm import Query
 from tuna.utils.logger import setup_logger
 from tuna.dbBase.sql_alchemy import DbSession
 from tuna.machine import Machine
+from tuna.utils.db_utility import get_solver_ids
 from tuna.utils.utility import check_qts
 from tuna.metadata import MYSQL_LOCK_WAIT_TIMEOUT, CONV_CONFIG_COLS, TENSOR_PRECISION
 from tuna.metadata import BN_DEFAULTS, BN_CONFIG_COLS
@@ -41,7 +42,6 @@ from tuna.metadata import FUSION_DEFAULTS, CONV_2D_DEFAULTS, CONV_3D_DEFAULTS
 from tuna.metadata import NUM_SQL_RETRIES
 from tuna.miopen_tables import TensorTable, Session
 from tuna.config_type import ConfigType
-from tuna.metadata import get_solver_ids
 
 LOGGER = setup_logger('helper')
